Threading doesn't work well with pygame's display.update().
 - Remove threading for that part of the code.
 - Make some more stub file changes.

+++ src/components/non_player_controlled.py
@@ -201,7 +201,7 @@ class Npc(pygame.sprite.Sprite):
         self.auto_walk()
 
 
-    def update(self, dt: float, collidable_group: pygame.sprite.Group) -> None:
+    def update(self, dt: int, collidable_group: pygame.sprite.Group) -> None:
         self.dt = dt
         self.collidable_group = collidable_group
